general: manage file events
Until now, any change in the desktop was managed by removing all the
items, reading all of them again, and painting all again.

This patch changes this, and now each event is managed individually,
by creating and removing only the specified file. This avoids the
flicker in the desktop each time an operation is done, and makes it
faster.

Anyway, if there is a problem when managing the file events (like
trying to get the attributes of a file that have been deleted too
fast), do a full read to synchronize again the file list. This
happens, for example, when creating a file with 'vim'.

To simplify the operations, changed the _fileList from a list into a
dictionary, to allow to find easily the files in events.