Concentrate data in the window itself
This patch modes the code to move the window to its place and to
the bottom into a function inside the window itself, to simplify
the code. Also, in case the window, when mapped, still doesn't
have the title set (currently needed to know if it is a desktop
window or not), this new code will defer the decission until the
title is set.