New method for removing the window from window list
Now it uses a different approach, based on overriding the getter and
the function used to get the value of the properties of a window.
With this code we cheat Gnome Shell in a more consistent, predictable
and future-proof way.

Still have to discover why it doesn't work with DashToDock