Fixed incorrect call to notify
The previous code used an incorrect call to notify, so, in fact,
it wasn't calling it. Also discovered that it should never call
notify on 'skip-taskbar', because gnome shell presumes that it
is triggered only if the value changed, and uses its value (true
or false) to increment or decrement a counter of "interesting"
windows (in C code). Since the code is tampering at javascript
level, it is not possible for the C code to read the desired
value, and decrements when it should increment.